requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
asyncer==0.0.7
attrs==24.2.0
bidict==0.23.1
cachetools==5.5.0
certifi==2024.8.30
chainlit==1.3.2
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gritql==0.1.5
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.12.0
httpcore==0.15.0
httplib2==0.22.0
httpx==0.23.0
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-cli==0.0.33
langchain-community==0.3.9
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-google-genai==2.0.6
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langserve==0.3.0
langsmith==0.1.147
Lazify==0.4.0
literalai==0.0.623
markdown-it-py==3.0.0
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.56.2
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.12
packaging==23.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
simple-websocket==1.1.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sse-starlette==1.8.2
starlette==0.41.3
syncer==2.0.3
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.28.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
psycopg2-binary==2.9.10
